Fix in situ insertion and refactor database package
