+++ air-quality-backend/scripts/run_in_situ_etl.py
@@ -2,10 +2,8 @@ from datetime import datetime, timedelta
 from dotenv import load_dotenv
 import logging
 from logging import config
-from src.database.air_quality_dashboard_dao import (
-    get_locations_by_type,
-    insert_data_openaq,
-)
+from src.database.locations import get_locations_by_type
+from src.database.in_situ import insert_data
 from src.etl.in_situ.openaq_dao import fetch_in_situ_measurements
 from src.etl.in_situ.openaq_adapter import transform_in_situ_data
 
@@ -29,7 +27,7 @@ def main():
     transformed_in_situ_data = transform_in_situ_data(in_situ_measurements_by_city)
 
     logging.info("Persisting in situ data")
-    insert_data_openaq(transformed_in_situ_data)
+    insert_data(transformed_in_situ_data)
 
 
 if __name__ == "__main__":
