+++ air-quality-backend/scripts/run_purge_old_data.py
@@ -2,9 +2,8 @@ from datetime import datetime, timedelta
 from dotenv import load_dotenv
 import logging
 from logging import config
-from src.database.air_quality_dashboard_dao import (
-    delete_data_before,
-)
+from src.database.forecasts import delete_data_before as delete_forecast_data
+from src.database.in_situ import delete_data_before as delete_in_situ_data
 
 config.fileConfig("../logging.ini")
 
@@ -15,7 +14,8 @@ def main():
     ten_days_previous = datetime.now() - timedelta(days=10)
 
     logging.info(f"Purging data earlier than {ten_days_previous}")
-    delete_data_before(ten_days_previous)
+    delete_forecast_data(ten_days_previous)
+    delete_in_situ_data(ten_days_previous)
 
 
 if __name__ == "__main__":
