--- air-quality-backend/src/database/location.py
@@ -1,8 +1,0 @@
-from typing import TypedDict
-
-
-class AirQualityLocation(TypedDict):
-    name: str
-    latitude: float
-    longitude: float
-    type: str
